Renames perma(-URL) to frozen(-URL)
In general, a perma-URL is one that is supposed to be valid fro a long time,
not one that points to a specific version of a file.
One could argue, that the later kind of constitutes a perma-URL,
but ... that is very sketchy arguing,
so lets better use an other word instead (frozen (in time)).