HUGE Refactor [major]
Too many things to list all; some:

- Thingiverse needs a few tiny fixes
- Wikifactory turned off their API (and generally seem half down) -> not supporting anymore!


I decided to move towards a sleeker scope,
where only the raw data will be fetched.
It will be stored in its own file
in the original format and with the original content,
and an extra file with crawling meta data will be put next to it.

These steps will be left for other, external tools:

- Format/OKH-version identification
- Sanitizing (replacing or removing invalid values or bad syntax)
- conversion to the current manifest format
- conversion to RDF